@@ -409,7 +409,7 @@ class EntityOperations implements ContainerInjectionInterface {
       case 'node':
         // Get the node entity to be flagged later.
         $flag_entity = $entity;
-        $flag_type = FlagType::FOLLOW_CONTENT;
+        $flag_type = $flag_entity->bundle() === 'event' ? FlagType::FOLLOW_EVENT : FlagType::FOLLOW_CONTENT;
         break;
 
     }
