@@ -363,7 +363,7 @@ class EICGroupsHelper implements EICGroupsHelperInterface {
       }
     }
     if ($entity) {
-      return $this->getGroupByEntity($entity);
+      return $this->getOwnerGroupByEntity($entity);
     }
     return FALSE;
   }
@@ -371,7 +371,7 @@ class EICGroupsHelper implements EICGroupsHelperInterface {
   /**
    * {@inheritdoc}
    */
-  public function getGroupByEntity(EntityInterface $entity) {
+  public function getOwnerGroupByEntity(EntityInterface $entity) {
     $group = FALSE;
     if ($entity instanceof GroupInterface) {
       return $entity;
@@ -901,7 +901,7 @@ class EICGroupsHelper implements EICGroupsHelperInterface {
           break;
         }
 
-        $group = $this->getGroupByEntity($node);
+        $group = $this->getOwnerGroupByEntity($node);
 
         if (!$group) {
           break;
@@ -977,7 +977,7 @@ class EICGroupsHelper implements EICGroupsHelperInterface {
     $countries_map = CountryManager::getStandardList();
     $location_formatted = $address[FieldHelper::getPropertyName(AddressField::ADDRESS_LINE1)]
       . ' ' .
-      $address[$address[FieldHelper::getPropertyName(AddressField::ADDRESS_LINE2)]] . '<br />';
+      $address[FieldHelper::getPropertyName(AddressField::ADDRESS_LINE2)] . '<br />';
     $location_formatted .= $address[FieldHelper::getPropertyName(AddressField::POSTAL_CODE)]
       . ' ' .
       $address[FieldHelper::getPropertyName(AddressField::LOCALITY)] . '<br />';
