@@ -4,35 +4,56 @@ namespace Drupal\eic_messages\Service;
 
 use Drupal\Core\DependencyInjection\ContainerInjectionInterface;
 use Drupal\Core\Entity\ContentEntityInterface;
+use Drupal\Core\Entity\EntityTypeManagerInterface;
 use Drupal\eic_groups\GroupsModerationHelper;
+use Drupal\eic_messages\Util\LogMessageTemplates;
 use Drupal\eic_user\UserHelper;
 use Symfony\Component\DependencyInjection\ContainerInterface;
 
 /**
- * Class GroupMessageCreator.
+ * Provides a message creator class for groups.
  */
 class GroupMessageCreator implements ContainerInjectionInterface {
 
   /**
+   * The message bus service.
+   *
    * @var \Drupal\eic_messages\Service\MessageBusInterface
    */
-  private $messageBus;
+  protected $messageBus;
 
   /**
+   * The EIC user helper service.
+   *
    * @var \Drupal\eic_user\UserHelper
    */
-  private $userHelper;
+  protected $userHelper;
 
   /**
+   * The entity type manager.
+   *
+   * @var \Drupal\Core\Entity\EntityTypeManagerInterface
+   */
+  protected $entityTypeManager;
+
+  /**
+   * Constructs new GroupMessageCreator object.
+   *
    * @param \Drupal\eic_messages\Service\MessageBusInterface $message_bus
+   *   The message bus service.
    * @param \Drupal\eic_user\UserHelper $user_helper
+   *   The EIC user helper service.
+   * @param \Drupal\Core\Entity\EntityTypeManagerInterface $entity_type_manager
+   *   The entity type manager.
    */
   public function __construct(
     MessageBusInterface $message_bus,
-    UserHelper $user_helper
+    UserHelper $user_helper,
+    EntityTypeManagerInterface $entity_type_manager
   ) {
     $this->messageBus = $message_bus;
     $this->userHelper = $user_helper;
+    $this->entityTypeManager = $entity_type_manager;
   }
 
   /**
@@ -41,7 +62,8 @@ class GroupMessageCreator implements ContainerInjectionInterface {
   public static function create(ContainerInterface $container) {
     return new static(
       $container->get('eic_messages.message_bus'),
-      $container->get('eic_user.helper')
+      $container->get('eic_user.helper'),
+      $container->get('entity_type.manager')
     );
   }
 
@@ -102,6 +124,17 @@ class GroupMessageCreator implements ContainerInjectionInterface {
         ]);
         break;
     }
+
+    // Create log message about the group state change.
+    $this->entityTypeManager->getStorage('message')
+      ->create([
+        'template' => LogMessageTemplates::GROUP_STATE_CHANGE,
+        'field_group_ref' => ['target_id' => $entity->id()],
+        'field_previous_moderation_state' => $original_state,
+        'field_moderation_state' => $current_state,
+        'uid' => $author_id,
+      ])
+      ->save();
   }
 
 }
