@@ -64,29 +64,15 @@ class GroupOverviewPages {
    *   The URL object or NULL if does not apply.
    */
   public static function getGroupOverviewPageUrl(string $page, GroupInterface $group) {
-    $page_id = NULL;
-    switch ($page) {
-      case 'discussions':
-        $page_id = self::DISCUSSIONS;
-        break;
-
-      case 'files':
-        $page_id = self::FILES;
-        break;
-
-      case 'events':
-        $page_id = self::EVENTS;
-        break;
-
-      case 'members':
-        $page_id = self::MEMBERS;
-        break;
-
-      case 'group_search':
-        $page_id = self::SEARCH;
-        break;
-
-    }
+    $pages = [
+      'discussions' => self::DISCUSSIONS,
+      'files' => self::FILES,
+      'events' => self::EVENTS,
+      'members' => self::MEMBERS,
+      'news' => self::NEWS,
+      'group_search' => self::SEARCH,
+    ];
+    $page_id = $pages[$page] ?? NULL;
 
     if (!empty($page_id) && !$group->isNew()) {
       return Url::fromRoute($page_id, ['group' => $group->id()]);
