@@ -8,6 +8,7 @@ import wiki from './wiki.data';
 import activity from './activity.data';
 import organisation from './organisation.data';
 import myActivity from './myActivity.data';
+import mySettings from './mySettings.data';
 
 export default {
   about,
@@ -19,5 +20,6 @@ export default {
   document,
   activity,
   organisation,
-  myActivity
+  myActivity,
+  mySettings
 };
