@@ -48,12 +48,22 @@ export default {
     date_time: '2018-06-24',
   },
   icon_file_path: common.icon_file_path,
-  type: {
-    extra_classes: "ecl-icon--currentColor",
-    icon: {
-      name: 'video',
-      type: 'general',
+  type: [
+    {
+      extra_classes: "ecl-icon--currentColor",
+      icon: {
+        name: 'video',
+        type: 'general',
+      },
+      label: 'Remote event',
     },
-    label: 'Remote event',
-  },
+    {
+      extra_classes: "ecl-icon--currentColor",
+      icon: {
+        name: 'map-marker',
+        type: 'custom',
+      },
+      label: 'Antwerp, Belgium',
+    }
+  ]
 };
