@@ -9,7 +9,10 @@ const element = document.getElementById('comments-discussion');
 ReactDOM.render(
   <>
     <ReactNotification />
-    <CommentsDiscussion discussionId={element.dataset.discussionId}/>
+    <CommentsDiscussion
+        discussionId={element.dataset.discussionId}
+        highlightedCommentId={element.dataset.highlightedComment}
+    />
   </>,
   element
 );
