@@ -29,6 +29,7 @@ class CommentsDiscussion extends React.Component {
 
     this.wrapperRef = React.createRef();
 
+
     this.fetchComments = this.fetchComments.bind(this);
     this.addComment = this.addComment.bind(this);
     this.handleClickOutside = this.handleClickOutside.bind(this);
@@ -47,6 +48,10 @@ class CommentsDiscussion extends React.Component {
   componentDidMount() {
     this.fetchComments();
     document.addEventListener('click', this.handleClickOutside);
+
+    if (this.wrapperRef.current && this.props.highlightedCommentId) {
+      document.getElementById('ecl-comment-overview__wrapper').scrollIntoView()
+    }
   }
 
   componentWillUnmount() {
@@ -172,6 +177,7 @@ class CommentsDiscussion extends React.Component {
     const params = {
       page: page || this.state.page,
       parentId: parentId,
+      highlightedComment: this.props.highlightedCommentId
     };
 
     axios.get(getPrefixDiscussionEndpoint() + '/' + this.props.discussionId + '/comments',
@@ -297,7 +303,7 @@ class CommentsDiscussion extends React.Component {
           size={50}
         />
       }
-      <div ref={this.wrapperRef} className="ecl-comment-overview__items">
+      <div ref={this.wrapperRef} id='ecl-comment-overview__wrapper' className="ecl-comment-overview__items">
         {this.state.comments.map(comment => {
           return <TopAnswer
             key={comment.comment_id}
