@@ -35,7 +35,7 @@ class Results extends React.Component {
 
   render() {
     if (!this.props.initiated) {
-      return <CircularProgress style={{top: '50%', left: '50%', position: 'fixed', zIndex: '51'}} className="loader" size={50} />
+      return <CircularProgress style={{ top: '50%', left: '50%', position: 'fixed', zIndex: '51' }} className="loader" size={50} />
     }
 
     if (Object.keys(this.props.results).length === 0 || (this.props.results.hasOwnProperty('numFound') && this.props.results.numFound === 0))
@@ -60,6 +60,7 @@ class Results extends React.Component {
               result={value}
               isGroupOwner={this.props.isGroupOwner}
               groupAdmins={this.props.groupAdmins}
+              type={this.props.bundle}
             />
           })}
         </div>
