@@ -3,14 +3,17 @@ require('laravel-mix-svg');
 
 // Public path is where the files will be created
 mix.setPublicPath(`dist/react`);
-
+mix.options({ runtimeChunkPath: 'custom' });
 mix.js(`react/components/Block/Overview/entrypoint.js`, `dist/react/block/overview`)
-mix.js(`react/components/Field/EntityTree/entrypoint.js`, `dist/react/field/entity_tree`)
-mix.js(`react/components/Block/ActivityStream/entrypoint.js`, `dist/react/block/activity_stream`)
-mix.js(`react/components/Block/CommentsDiscussion/entrypoint.js`, `dist/react/block/comments_discussion`)
-mix.js(`react/components/Block/ShareModal/entrypoint.js`, `dist/react/block/share_modal`)
-mix.js(`react/components/Block/Gallery/entrypoint.js`, `dist/react/block/gallery`)
-mix.js(`react/components/Block/Announcements/entrypoint.js`, `dist/react/block/announcements`)
+.js(`react/components/Field/EntityTree/entrypoint.js`, `dist/react/field/entity_tree`)
+.js(`react/components/Block/ActivityStream/entrypoint.js`, `dist/react/block/activity_stream`)
+.js(`react/components/Block/CommentsDiscussion/entrypoint.js`, `dist/react/block/comments_discussion`)
+.js(`react/components/Block/ShareModal/entrypoint.js`, `dist/react/block/share_modal`)
+.js(`react/components/Block/Gallery/entrypoint.js`, `dist/react/block/gallery`)
+.js(`react/components/Block/Announcements/entrypoint.js`, `dist/react/block/announcements`)
+.js(`react/components/Block/NotifManagement/entrypoint.js`, `dist/react/block/notifmanagement`)
+.js(`react/components/Block/Toggle/entrypoint.js`, `dist/react/block/toggle`)
+.extract(['react', 'react-dom', 'react-router-dom', 'axios', '@material-ui', 'lodash-es', 'scheduler'], 'dist/react/vendor.js');
 
 mix.svg({
   class: 'ecl-icon',
